+++ sky-server/src/main/java/com/sky/service/DishService.java
@@ -54,5 +54,5 @@ public interface DishService {
      * @param categoryId
      * @return
      */
-    List<Dish> getByCategoryId(Integer categoryId);
+    List<Dish> getByCategoryId(Long categoryId);
 }
